[ADD] apple login response 정리(#350)
